// src/interfaces/sales.interface.ts
export interface SalesInterface {
  date: string;
  units: string;
  units_lastyear: string;
  revenue: string;
  revenue_lastyear: string;
  cogs: string;
  cogs_lastyear: string;
  ads_cost: string;
  ads_cost_lastyear: string;
  best_sellers_rank: string;
  best_sellers_rank_lastyear: string;
  ppc_revenue_adjusted: string;
  ppc_revenue_adjusted_lastyear: string;
  ads_cost_adjusted: string;
  ads_cost_adjusted_lastyear: string;
  units_dd: string;
  revenue_dd: string;
  cogs_dd: string;
  ads_dd: string;
  best_sellers_rank_dd: string;
  units_projected: string;
  revenue_projected: string;
  cogs_projected: string;
  ads_cost_projected: string;
  best_sellers_rank_projected: string;
  ppc_revenue_projected: string;
  ppc_revenue: string;
  ppc_revenue_lastyear: string;
}